x/rio/types/message_create_cert_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"rio/testutil/sample"
)

func TestMsgCreateCert_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateCert
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateCert{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateCert{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
